Use iterables for text_from_bytes.
We’re gonna need this to be a generator function, and it’s too much of a
pain to write code that dynamically figures out if it’s getting just one
value or a list of values.  It’s our code, so just assume we’re always
passing a list / iterable, and then make sure we do, in fact, always
pass a list / iterable.