Fudge the movie header to accommodate all inputs.
There’s a check in Dolphin’s Movie::CheckInputEnd() that stops the movie
if it hits any of 3 counts in the movie header.  The counts are
- frames
- inputs
- clock ticks

We know the exact number of inputs, and quick experiments suggest there
are about 4 inputs per frame and about 2.025M ticks per input.

Fudge some overestimates of frames and ticks, so that the movie always
runs to the end of all inputs.