Use just one input file (default stdin) from the command line.
The old API allowed as many input files as the user wanted, with a
hyphen to signify reading from stdin.  It seemed kinda convenient at
design time, but now that I’m trying to fit this program into other bash
scripts (and especially Makefiles) it’s getting to be a bit of a pain to
craft just the right list of input files / stdin sigil.

Better solution is to split the concerns.  Let the shell or whatever
manage building one giant list of inputs / feed a whole bunch of stuff
to stdin in the proper order.  Let this program just worry about reading
from an input stream and doing the right thing.